[1.0] release time! :3
updates:
- the plugin now checks if there is an update available for download
- this check can be disabled by starting OMSI with the `-omsipresence_noupdate` launch option

debug:
- debug mode is no longer a compile-only define: start OMSI with the `-omsipresence_debug` launch option to activate
- logging is no longer a compile-only define: logs are automatically added to `logfile.txt`

discord:
- local strings in `Update()` are now stored on the stack instead of the heap (idk why i did that in the first place lol)

misc:
- various code, style and comment rewrites/cleanup (TODO: reorganize headers?)
- removed unused helper functions
- slight README rewrite

+++ OMSIPresence/src/veh.cpp
@@ -1,6 +1,6 @@
 #include "veh.h"
 
-void veh::AddHandler()
+void VEH::AddHandler()
 {
 	if (!handler)
 	{
@@ -8,10 +8,10 @@ void veh::AddHandler()
 		return;
 	}
 
-	DEBUG(dbg::error, "Tried to veh::AddHandler() when the handler already exists");
+	Log(LT_ERROR, "Tried to veh::AddHandler() when the handler already exists");
 }
 
-void veh::RemoveHandler()
+void VEH::RemoveHandler()
 {
 	if (handler)
 	{
@@ -20,12 +20,14 @@ void veh::RemoveHandler()
 		return;
 	}
 
-	DEBUG(dbg::error, "Tried to veh::RemoveHandler() when the handler doesn't exist");
+	Log(LT_ERROR, "Tried to veh::RemoveHandler() when the handler doesn't exist");
 }
 
-LONG CALLBACK veh::ExceptionHandler(EXCEPTION_POINTERS* exception_pointers)
+// Exceptions in our code aren't the *worst* thing to happen to us (OMSI has them all the time), but there should be zero tolerance for something so simple to break
+// Since we're entering a heavily unwanted state, our exception handler creates a dump and either terminates the game, or just adds a log entry if we're in debug
+LONG CALLBACK VEH::ExceptionHandler(EXCEPTION_POINTERS* exception_pointers)
 {
-	veh::RemoveHandler();
+	VEH::RemoveHandler();
 
 	// Set up minidump information
 	MINIDUMP_EXCEPTION_INFORMATION exception_info;
@@ -48,7 +50,7 @@ LONG CALLBACK veh::ExceptionHandler(EXCEPTION_POINTERS* exception_pointers)
 
 	CloseHandle(hFile);
 
-	// Get the name of the module our address is in
+	// Try to get the name of the module our address is in
 	PVOID address = exception_pointers->ExceptionRecord->ExceptionAddress;
 
 	char module_name[MAX_PATH] = { 0 };
@@ -61,21 +63,23 @@ LONG CALLBACK veh::ExceptionHandler(EXCEPTION_POINTERS* exception_pointers)
 		}
 	}
 
-	// If we're in release, error and terminate. If we're in debug, just print a log message
 	DWORD code = exception_pointers->ExceptionRecord->ExceptionCode;
 
-#ifndef PROJECT_DEBUG
-	Error("Exception %08X at %08X%s. OMSI 2 cannot continue safely and will be terminated.\n\n"
-		"An %s file containing more information about the crash has been created in your game folder. "
-		"Please submit this file for developer review.",
-		code, address, module_name, dmp_filename);
+	if (debug)
+	{
+		Log(LT_ERROR, "Exception %08X at %08X%s. Saved to %s", code, address, module_name, dmp_filename);
+	}
+	else
+	{
+		Error("Exception %08X at %08X%s. OMSI 2 cannot continue safely and will be terminated.\n\n"
+			"An %s file containing more information about the crash has been created in your game folder. "
+			"Please submit this file for developer review.",
+			code, address, module_name, dmp_filename);
 
-	// Normally, we would pass this exception on using EXCEPTION_EXECUTE_HANDLER in our exception handlers which call this function
-	// Unfortunately, the game's built-in exception handler will throw it away, so we'll just have to terminate the program here
-	TerminateProcess(hProcess, ERROR_UNHANDLED_EXCEPTION);
-#else
-	DEBUG(dbg::error, "Exception %08X at %08X%s. Saved to %s", code, address, module_name, dmp_filename);
-#endif
+		// Normally, we would pass this exception on using EXCEPTION_EXECUTE_HANDLER in our exception handlers which call this function
+		// Unfortunately, the game's built-in exception handler will throw it away, so we'll just have to terminate the program here
+		TerminateProcess(hProcess, ERROR_UNHANDLED_EXCEPTION);
+	}
 
 	return EXCEPTION_EXECUTE_HANDLER;
 }
